feat: Implement Calendar Data Extraction and Visibility Management Services
- Added CalendarDataExtractor service for extracting calendar data from DOM elements, including email, name, and visibility.
- Implemented CalendarVisibilityManager service to handle visibility changes for calendars, optimizing batch operations.
- Created VirtualScrollHandler service to manage virtual scrolling behavior for discovering all calendars.
- Developed integration tests for GoogleCalendarRepository to ensure proper orchestration of services and functionality.
- Added comprehensive unit tests for CalendarDataExtractor to validate email extraction logic and edge cases.
- Enhanced logging throughout services for better traceability and debugging.
